Clean up
Removed all the calendar test code into new activity
